requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
feedparser==6.0.11
anthropic==0.84.0
resend==2.26.0
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
lxml==5.3.0
